remove some hpp files
