+++ source/source_lcao/CMakeLists.txt
@@ -9,21 +9,28 @@ if(ENABLE_LCAO)
         hamilt_lcao.cpp
         module_operator_lcao/operator_lcao.cpp
         module_operator_lcao/veff_lcao.cpp
+        module_operator_lcao/veff_dh.cpp
         module_operator_lcao/meta_lcao.cpp
         module_operator_lcao/op_dftu_lcao.cpp
         module_operator_lcao/deepks_lcao.cpp
         module_operator_lcao/op_exx_lcao.cpp
         module_operator_lcao/overlap.cpp
+        module_operator_lcao/overlap_fs.cpp
         module_operator_lcao/ekinetic.cpp
+        module_operator_lcao/ekinetic_fs.cpp
+        module_operator_lcao/ekinetic_dh.cpp
         module_operator_lcao/nonlocal.cpp
         module_operator_lcao/nonlocal_dh.cpp
         module_operator_lcao/nonlocal_fs.cpp
         module_operator_lcao/td_ekinetic_lcao.cpp
         module_operator_lcao/td_nonlocal_lcao.cpp
         module_operator_lcao/td_pot_hybrid.cpp
+        module_operator_lcao/td_pot_hybrid_fs.cpp
         module_operator_lcao/dspin_lcao.cpp
+        module_operator_lcao/dspin_fs.cpp
         module_operator_lcao/dftu_lcao.cpp
-        module_operator_lcao/operator_force_stress_utils.cpp
+        module_operator_lcao/dftu_fs.cpp
+        module_operator_lcao/operator_fs_utils.cpp
         dftu_lcao.cpp
         pulay_fs_center2.cpp
         FORCE_STRESS.cpp
