+++ source/source_lcao/module_operator_lcao/operator_fs_utils.hpp
@@ -1,7 +1,7 @@
-#ifndef OPERATOR_FORCE_STRESS_UTILS_HPP
-#define OPERATOR_FORCE_STRESS_UTILS_HPP
+#ifndef OPERATOR_FS_UTILS_HPP
+#define OPERATOR_FS_UTILS_HPP
 
-#include "operator_force_stress_utils.h"
+#include "operator_fs_utils.h"
 #include "source_base/parallel_reduce.h"
 #include "source_base/timer.h"
 #include "source_cell/module_neighbor/sltk_grid_driver.h"
@@ -193,4 +193,4 @@ void cal_force_stress_2center(
 
 } // namespace OperatorForceStress
 
-#endif // OPERATOR_FORCE_STRESS_UTILS_HPP
+#endif // OPERATOR_FS_UTILS_HPP
